Allow plugins to override the QML VideoOutput type.
Move QDeclarativeVideoOutput to the private QtMultimediaQuickTools
library to make the QDeclarativeVideoOutputBackend interface
implementable by a plugin.  It also adds  private control for the
gstreamer backend which allows a custom sink to be specified instead
of one of the defaults.

This provides a (use at your own risk) option to inject a non-standard
gstreamer video sink element into the pipeline.

Change-Id: I763c483a1fc9ec56dc7b8be0bc71523f029a36ee